refactor(common): se agregan mas verificaciones
Se agregan verificaciones para el tamanio del nombre del path y del archivo.
No se permite la creacion de archivos en niveles mayores al MAX_LEVEL establecido

+++ fisopfs/bitmap.h
@@ -26,6 +26,6 @@ void clear_bit(bitmap_t *bitmap, int n);
 
 int get_bit(bitmap_t *bitmap, int n);
 
-long get_free_bit(bitmap_t *bitmap);
+int get_free_bit(bitmap_t *bitmap);
 
 #endif  // _BITMAP_H_
